requirements.txt
numpy
scipy
pandas
h5py
scikit-learn
sklearn-pandas
pandas-summary
imbalanced-learn
fancyimpute
matplotlib
seaborn
cython
jupyter
jupyter_contrib_nbextensions
tensorflow
keras
tqdm
xgboost